refactor(main): browser
