+++ src/main.browser.ts
@@ -1,0 +1,69 @@
+/*
+ * Providers provided by Angular
+ */
+import {bootstrap} from 'angular2/platform/browser';
+import {provide, enableProdMode} from 'angular2/core';
+
+/*
+ * App Component
+ * our top level component that holds all of our components
+ */
+import {DIRECTIVES, PIPES, PROVIDERS} from './platform/browser';
+import {App} from './app/app';
+
+
+if ('production' === ENV) {
+  // Production
+  enableProdMode();
+} else {
+  // Development
+
+}
+
+/*
+ * Bootstrap our Angular app with a top level component `App` and inject
+ * our Services and Providers into Angular's dependency injection
+ */
+export function main() {
+  return bootstrap(App, [
+    ...PROVIDERS,
+    ...PIPES,
+    ...PROVIDERS
+  ])
+  .catch(err => console.error(err));
+}
+
+
+/*
+ * Vendors
+ * For vendors for example jQuery, Lodash, angular2-jwt just import them anywhere in your app
+ * You can also import them in vendors to ensure that they are bundled in one file
+ * Also see custom-typings.d.ts as you also need to do `typings install x` where `x` is your module
+ */
+
+
+/*
+ * Hot Module Reload
+ * experimental version by @gdi2290
+ */
+
+function bootstrapDomReady() {
+  // bootstrap after document is ready
+  return document.addEventListener('DOMContentLoaded', main);
+}
+
+if ('development' === ENV) {
+  // activate hot module reload
+  if (HMR) {
+    if (document.readyState === 'complete') {
+      main();
+    } else {
+      bootstrapDomReady();
+    }
+    module.hot.accept();
+  } else {
+    bootstrapDomReady();
+  }
+} else {
+  bootstrapDomReady();
+}
